started PID motor movement

+++ ArduinoSide/Arduino Modularized Code/main_final/PID.h
@@ -1,1 +1,1 @@
-float PID(float error, float &pastError, float &integral, float kp, float ki, float kd);
+double PID(double error, double &pastError, double &integral, double kp, double ki, double kd);
